refactor: extract status bar tests into statusBarSuite.ts
